refactor/#73: direcotry 수정
