+++ src/features/attachment/attachment.module.ts
@@ -5,7 +5,7 @@ import { AttachmentRepository } from '@features/attachment/repositories/attachme
 import { AttachmentMapper } from '@features/attachment/mappers/attachment.mapper';
 import { NestjsFormDataModule } from 'nestjs-form-data';
 import { S3Module } from '@libs/s3/s3.module';
-import { CreateAttachmentsCommandHandler } from '@features/attachment/commands/create-user/create-attachment.command-handler';
+import { CreateAttachmentsCommandHandler } from '@features/attachment/commands/create-attachment/create-attachment.command-handler';
 
 const controllers = [AttachmentController];
